Use side effects to determine whether we should run the node or not
